% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmojiViz.R
\name{EmojiViz}
\alias{EmojiViz}
\title{Basic WhatsApp Chatlog Statistics}
\usage{
EmojiViz(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  min.occur = 1,
  return.data = FALSE,
  EmojiVec = "all",
  plot = "bar"
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with WhatsAppParse()}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{min.occur}{Minimum number of occurances for Emoji to be included in the plots. Default is 1.}

\item{return.data}{If TRUE, returns the subsetted dataframe. Default is FALSE.}

\item{EmojiVec}{A vector of Emoji that the visualizations will be restricted to}

\item{plot}{The type of plot that should be outputted. Options include "heatmap", "cumsum", "bar" and "splitbar"}
}
\value{
Plots and/or the subsetted dataframe based on author names, datetime and Emoji occurance
}
\description{
Creates a list of basic information about a single WhatsApp chatlog
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
EmojiViz(data)
}
