% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeDiPlo.R
\name{LeDiPlo}
\alias{LeDiPlo}
\title{Lexical Disperson plots for Keywords in WhatsApp chatlogs}
\usage{
LeDiPlo(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  palettes = "Paired",
  keywords = c("hello", "world")
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with WhatsAppParse()}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{palettes}{Palette name to be used. Default is "Paired"}

\item{keywords}{A vector of keywords to be displayed, default is c("hello",world")}
}
\value{
Lexical Disperson plots for specified keywords
}
\description{
Visualizes the occurance of specific keywords within the chat
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
LeDiPlo(data, keywords = c("smilies","handy"))
}
