% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chat.R
\name{parse_chat}
\alias{parse_chat}
\title{Parsing exported WhatsApp chatlogs as a dataframe}
\usage{
parse_chat(
  path,
  os = "auto",
  language = "auto",
  anonimize = "add",
  consent = NA,
  emoji_dictionary = "internal",
  smilie_dictionary = "wikipedia",
  order = "both",
  rpnl = " start_newline ",
  rpom = " media_omitted "
)
}
\arguments{
\item{path}{Character string containing the file path to the exported WhatsApp chatlog.}

\item{os}{Operating system of the phone the chat was exported from. Default "auto" tries to automatically detect the OS. OS also supports "android" or "iOS".}

\item{language}{Indicates the language setting of the phone with which the messages were exported. Default is "auto" trying to match either 'English' or 'German'. More languages might be supported in the future.}

\item{anonimize}{TRUE results in the vector of sender names being anonymized and columns containing personal identifiable information to be deleted, FALSE displays the actual names and all content, "add" adds
anonomized columns to the full info cilumns. Do not blindly trust this and always double check.}

\item{consent}{String containing a consent message. All messages from users who have not posted this exact message into the chat will be deleted. Default is NA.}

\item{emoji_dictionary}{Dictionary for emoji matching. Can use a version included in this package when set to "internal" or
an updated data frame created by \code{\link[WhatsR]{download_emoji}}.}

\item{smilie_dictionary}{"emoticons" uses \code{\link[qdapRegex]{ex_emoticon}} to extract smilies, "wikipedia" uses a more inclusive custom list
of smilies containing all mentions from https://de.wiktionary.org/w/index.php?title=Verzeichnis:International/Smileys
and manually added ones.}

\item{order}{Can be "time" or "both". Controls whether an indicator column for the order of messages is added
"time" orders the messages according to the WhatsApp Timestamp the message received while it was sent.
"both" gives two columns with the respective orders when the message was sent and how it was displayed on the exporting phone.}

\item{rpnl}{Replace newline. A character string for replacing line breaks within messages for the parsed message for better readability. Default is " start_newline ".}

\item{rpom}{Replace omitted media. A character string replacing the indicator for omitted media files for better readability. Default is " media_omitted ".}
}
\value{
For anonimize = "add", returns a data frame containing:

     1) A column to indicate the date and time when the message was sent \cr
     2) A column containing the anonymized name of the sender \cr
     3) A column to indicate the name of the sender \cr
     4) A column containing the raw message \cr
     5) A column containing a "flat" message, stripped of emoji, numbers, special characters, file attachments, sent locations etc. \cr
     6) A column containing a tokenized version of the flat message \cr
     7) A column containing only URLs that were contained in the messages \cr
     8) A column containing only the file types of attached media files (if any are contained)\cr
     9) A column containing only sent locations and indicators for shared live locations \cr
     10) A column containing only emoji that were used in the message \cr
     11) A column containing only textual descriptions of emoji that were used in the message \cr
     12) A column containing only emoticons (e.g. ":-)") that were used in the message \cr
     13) A column containing the number of tokens per message, derived from the "flattened" message \cr
     14) A column containing WhatsApp system messages in group chats (e.g."You added Frank to the group") \cr
     15) A column specifying the order of the rows according to the time stamp the messages have on the phone used for extracting the chatlog \cr
     16) A column for specifying the order of the rows as they are displayed on the phone used for extracting the chatlog \cr
}
\description{
Creates a dataframe from an exported WhatsApp chatlog containing one row per message. Some columns
are saved as lists using the I() function so that multiple elements can be stored per message while still maintaining
the general structure of one row per message. These columns should be treated as lists or unlisted first.
}
\examples{
data <- parse_chat(system.file("englishandroid24h.txt", package = "WhatsR"))
}
