% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseIos.R
\name{ParseIos}
\alias{ParseIos}
\title{Parsing raw WhatsApp Chat Textfiles accoring to iOs text structure}
\usage{
ParseIos(
  UnparsedChat,
  nl = "\\n",
  mediaomitted = "<media omitted>",
  mediaindicator = "^<attached:\\\\s(.)*?\\\\.(.)*?>$",
  sentlocation = paste0("Location: (?=https:\\\\/\\\\/maps\\\\.google\\\\.com\\\\/",
    "\\\\?q=\\\\d\\\\d.\\\\d{6}\\\\,\\\\d\\\\.\\\\d{6})"),
  livelocation = "^live location shared$",
 
    datetimeindicator = "(?!^)(?=\\\\[((\\\\d{2}\\\\.\\\\d{2}\\\\.\\\\d{2})|(\\\\d{1,2}\\\\/\\\\d{1,2}\\\\/\\\\d{2})),\\\\s\\\\d{1,2}\\\\:\\\\d{2}((\\\\:\\\\d{2}\\\\s(?i:(pm|am)))|(\\\\s(?i:(pm|am)))|(\\\\:\\\\d{2}\\\\])|(\\\\:\\\\d{2})|(\\\\s))\\\\])",
  nlreplace = " start_newline ",
  mediareplace = " media_omitted "
)
}
\arguments{
\item{UnparsedChat}{the name of the exported Whatsapp textfile to be parsed as a character string}

\item{nl}{character string defining character for newline indicators. Default is a unicode newline}

\item{mediaomitted}{character string inserted by WhatsApp instead of filenames when not exporting media}

\item{mediaindicator}{character string for detecting media and file attachments}

\item{sentlocation}{Regex for detecting autogenerated messages for locations shared via chat}

\item{livelocation}{Regex for detecting autogenerated messages for locations shared via chat}

\item{datetimeindicator}{Regex for detecting the DateTime indicator at the beginning of each message}

\item{nlreplace}{replacement string for a newline character in parsed message. Default is " START_NEWLINE "}

\item{mediareplace}{replacement string for omitted media files. Default is " MEDIA_OMITTED "}
}
\value{
A dataframe containing the timestamp, name of the sender and message body
}
\description{
Creates a dataframe from an exported WhatsApp textfile containing one row per message
and a column for DateTime when the message was send, name of the sender and body of the message.
}
\examples{
ParsedChat <- ParseIos("[29.01.18, 23:33:00] Alice: Hello?\\\\n [29.01.18, 23:45:01] Bob: Hello")
}
