% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lexical_dispersion.R
\name{plot_lexical_dispersion}
\alias{plot_lexical_dispersion}
\title{Lexical Disperson plots for Keywords in WhatsApp chatlogs}
\usage{
plot_lexical_dispersion(
  data,
  names = "all",
  names_col = "Sender",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  keywords = c("hello", "world"),
  return_data = FALSE,
  exclude_sm = FALSE,
  ...
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with \code{\link[WhatsR]{parse_chat}}.}

\item{names}{A vector of author names that the plots will be restricted to.}

\item{names_col}{A column indicated by a string that should be accessed to determine the names. Only needs to be changed when \code{\link[WhatsR]{parse_chat}} used the parameter anon = "add" and the column "Anonymous" should be used. Default is "Sender".}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with \code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with \code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{keywords}{A vector of keywords to be displayed, default is c("hello",world").}

\item{return_data}{Default is FALSE, returns data frame of plot when TRUE.}

\item{exclude_sm}{If TRUE, excludes the WhatsApp System Messages from the descriptive statistics. Default is FALSE.}

\item{...}{Further arguments passed down to \code{\link[qdap]{dispersion_plot}}.}
}
\value{
Lexical Dispersion plots for specified keywords
}
\description{
Visualizes the occurrence of specific keywords within the chat
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_lexical_dispersion(data, keywords = c("auch"))
}
