% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_location.R
\name{plot_location}
\alias{plot_location}
\title{Plotting Locations sent in WhatsApp chatlogs on maps}
\usage{
plot_location(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  mapzoom = 5,
  return.data = FALSE,
  add.jitter = FALSE,
  jitter.val = 0.01,
  jitter.seed = 123,
  mapleeway = 0.1
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with code{\link[WhatsR]{parse_chat}}}

\item{names}{A vector of author names that the plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{mapzoom}{Value for zoom into the map passed down to get_map. default value is 5. Higher zoom will auto-download more mapfiles which can take a while.}

\item{return.data}{If TRUE, returns a dataframe of LatLon coordinates extracted from the chat for more elaborate plotting. Default is FALSE.}

\item{add.jitter}{IF TRUE, adds some random jitter to geolocations to obscure exact locations}

\item{jitter.val}{Amount of random jitter to add to the geolocations to hide exact locations. Default value is 0.01}

\item{jitter.seed}{Add}

\item{mapleeway}{Adds additional space to the map so that points do not sit exactly at the border of the plot. Default value is 5}
}
\value{
Plots for geolocation and/or a dataframe of LatLon coordinates
}
\description{
Plots the location data that is sent in the WhatsApp Cchatlog on an autoscaled map
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_location(data,mapzoom = 10)
}
