% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Visualize the Network of replies in Whatsapp chatlogs}
\usage{
plot_network(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  return.data = FALSE,
  collapse_sessions = FALSE,
  edgetype = "n"
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with \code{\link[WhatsR]{parse_chat}}}

\item{names}{A vector of author names that the visulaization will be restricted to. Non-listed authors will be removed}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Input is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Input is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{return.data}{If TRUE, returns a dataframe of subsequent interactions with senders and recipients. Default is FALSE.}

\item{collapse_sessions}{Whether multiple subsequent messages by the same sender should be collapsed into one row. Default is FALSE.}

\item{edgetype}{What type of content is displayed as an edge. Must be one of "TokCount","EmojiCount","SmilieCount","LocationCount","URLCount","MediaCount" or "n".}
}
\value{
A network visualization of authors in WhatsApp chatlogs wher each subsequent message is considered a reply to the previous one.
}
\description{
Plots a network for replies between authors in chatlogs
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_network(data)
}
