% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smilies.R
\name{plot_smilies}
\alias{plot_smilies}
\title{Visualize Smilies used in Whatsapp chatlogs}
\usage{
plot_smilies(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  min.occur = 1,
  return.data = FALSE,
  SmilieVec = "all",
  plot = "bar"
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with code{\link[WhatsR]{parse_chat}}}

\item{names}{A vector of author names that the plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{min.occur}{The minimum number of occurances a Smiley has to have to be included in the Visualization. Default is 1}

\item{return.data}{If TRUE, returns a dataframe of LatLon coordinates extracted from the chat for more elaborate plotting. Default is FALSE.}

\item{SmilieVec}{A vector of Smilies that the visualizations will be restricted to.}

\item{plot}{The type of plot that should be outputted. Options include "heatmap", "cumsum", "bar" and "splitbar"}
}
\value{
Plots for Replytimes of Reactiontimes of authors
}
\description{
Plots the emilies used in WhatsApp chatlogs
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_smilies(data)
}
