% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Wordclouds for WhatsApp chatlogs}
\usage{
plot_wordcloud(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  stop = "german",
  maxwords = 100,
  mincount = 1,
  comparison = FALSE
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with parse_chat()}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{stop}{The language of the chat for stopword removal, passed down to quanteda::stopwords(). default is "german"}

\item{maxwords}{Maximum number of words to display in the wordcloud. Default is 100.}

\item{mincount}{Minimum number of occurances a word has to have to be included in the wordplot. Default is 1}

\item{comparison}{If TRUE, compares the unique wordclouds for different authors for a maximum of 8 authors. Default is FALSE}
}
\value{
A wordcloud plot per author for WhatsApp chatlogs
}
\description{
Creates awordcloud by author for WhatsApp chatlogs
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_wordcloud(data, comparison = TRUE)
}
