% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceEmoji.R
\name{ReplaceEmoji}
\alias{ReplaceEmoji}
\title{Replacing Emoji with Text or Unicode}
\usage{
ReplaceEmoji(
  string,
  replacement = "text",
  start = " |Emoji_",
  end = "| ",
  EmojiDictionary = "internal",
  ...
)
}
\arguments{
\item{string}{a character string containing the Emojis to be replaced}

\item{replacement}{"text" replaces the character string with a textual descirption, "uni" with correpsonding unicode}

\item{start}{a character string for pasted to the beginning of the replacement, default is " |Emoji_"}

\item{end}{charactaer string pasted to the end of the replacement, default is "| "}

\item{EmojiDictionary}{Dictionary for Emoji matching. Can use a version included in this package when set to "interal" or
an updated dataframe created by \code{\link[WhatsR]{download_emoji}}}

\item{...}{arguments passed from other functions}
}
\value{
a character string containing textual descriptions or unicode instead of emoji
}
\description{
Replaces all instances of Emoji in a textstring with a textual description or their unicode equivalent
}
\examples{
example <- readRDS(system.file("EmojiExample.rds", package = "WhatsR"))
print(example)
Replacement <- ReplaceEmoji(example)
print(Replacement)
}
