% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TailorData.R
\name{Tailor}
\alias{Tailor}
\title{Basic WhatsApp Chatlog Statistics}
\usage{
Tailor(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  excludeSM = TRUE
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with WhatsAppParse()}

\item{names}{If TRUE, excludes the WhatsAppSystemmessages from the descriptive statistics. Default is TRUE.}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{excludeSM}{If TRUE, excludes the WhatsAppSystemmessages from the descriptive statistics. Default is TRUE.}
}
\value{
A dataframe that is restricted to the specified timeframe and authors
}
\description{
Creates a list of basic information about a single WhatsApp chatlog
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
Tailor(data, names = c("Mallory","Alice"))
}
