% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_chat.R
\name{tailor_chat}
\alias{tailor_chat}
\title{Restricting chat logs to certain authors or timeframes.}
\usage{
tailor_chat(
  data,
  names = "all",
  starttime = "1960-01-01 00:00",
  endtime = as.character(as.POSIXct(Sys.time(), tz = "UTC")),
  exclude_sm = FALSE
)
}
\arguments{
\item{data}{A WhatsApp chat log that was parsed with \code{\link[WhatsR]{parse_chat}}.}

\item{names}{A vector of names that the output is restricted to. Messages from other non-contained authors are excluded.}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with \code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm". Is interpreted as UTC to be compatible with WhatsApp timestamps.}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with \code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm". Is interpreted as UTC to be compatible with WhatsApp timestamps.}

\item{exclude_sm}{If TRUE, excludes the WhatsApp system messages from the descriptive statistics. Default is FALSE.}
}
\value{
A dataframe that is restricted to the specified timeframe and authors
}
\description{
Excluding parts of the chat by senders or timestamps
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
tailor_chat(data, names = c("Mallory", "Alice"))
}
