% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokensOverTime.R
\name{TokensOverTime}
\alias{TokensOverTime}
\title{Distribution of Tokens over time}
\usage{
TokensOverTime(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  plot = "alltime"
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with WhatsAppParse()}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{plot}{Time resolution for plots. Options include "year", "month", "day", "hour", "heatmap" and "alltime". Default is "alltime".}
}
\value{
A summary of tokens over time
}
\description{
Summarizes the distribution of tokens over time
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
TokensOverTime(data)
}
