% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkVisualization.R
\name{WaNetworkAnimation}
\alias{WaNetworkAnimation}
\title{Visualize the Network of replies in Whatsapp chatlogs}
\usage{
WaNetworkAnimation(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  return.data = FALSE,
  save = "html",
  filename = "whatsapp_network"
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with WhatsAppParse()}

\item{names}{A vector of author names that the visulaization will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{return.data}{If TRUE, returns a dataframe of LatLon coordinates extracted from the chat for more elaborate plotting. Default is FALSE.}

\item{save}{If FALSE, does not save the network visualization in a file. Any other value saves the visualization as an html file.}

\item{filename}{filename for the saved network visualization, default is "whatsapp_network"}
}
\value{
html file with network visualization of authors in WhatsApp chatlog
}
\description{
Plots a network for replies between authors in chatlogs
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
WaNetworkAnimation(data)
}
