% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chat.R
\name{parse_chat}
\alias{parse_chat}
\title{Parsing exported WhatsApp Textfiles as a Dataframe}
\usage{
parse_chat(
  name,
  EmojiDic = "internal",
  smilies = 2,
  anon = "add",
  media = TRUE,
  web = "domain",
  order = "both",
  language = "auto",
  os = "auto",
  rpnl = " start_newline ",
  rpom = " media_omitted ",
  consent = NA,
  ...
)
}
\arguments{
\item{name}{the name of the exported Whatsapp textfile to be parsed as a character string}

\item{EmojiDic}{Dictionary for Emoji matching. Can use a version included in this package when set to "interal" or
an updated dataframe created by \code{\link[WhatsR]{download_emoji}}}

\item{smilies}{1 uses \code{\link[qdapRegex]{ex_emoticon}} to extract smilies, 2 uses a more inclusive custom list
of smilies containing all mentions from https://de.wiktionary.org/w/index.php?title=Verzeichnis:International/Smileys
and manually added ones}

\item{anon}{TRUE results in the vector of sender names being anonymized, FALSE displays the actual names, "add" adds a
column of anomyized names next to the actual names}

\item{media}{TRUE/FALSE indicates whether the chatlog was downloaded with or without mediafiles. If TRUE, names of
attached mediafiles will be extracted into a seperate column.}

\item{web}{"domain" will shorten sent links to domains, "url" will display the full URL}

\item{order}{Can be "time" or "both". Whether an indicator column for the order of messages is added
that the messages were exported from. "time" orders the messages according to the WhatsApp Timestamp the message received while it was sent.
Due to internet problems, these orders are not necessarily interchangeable. "both" gives two columns with the respective orders}

\item{language}{Indicates the language setting of the phone with which the messages were exported. This is important because
it changes the structure of date/time columns and indicators for sent media. Default is "auto" trying to match either English or German. More languages might be supported in the future.}

\item{os}{operating system of the phone the chat was exported from. Default "auto" tries to automatically detect the os. Also manually upports "android" or "iOS"}

\item{rpnl}{Replace newline. A character string for replacing linebreaks within messages for the parsed message for better readibility. Default is " start_newline "}

\item{rpom}{Replace omitted media. A character string replacing the indicator for omitted media files for better readibility. Default is " media_omitted "}

\item{consent}{String containng a consent message. All messages from users who have not posted this exact message into the chat will be deleted. Default is NA.}

\item{...}{Further arguments passed down to ReplaceEmoji()}
}
\value{
A dataframe containing:

     1) A column to indicate the date and time when the message was send \cr
     2) A column containing the anonimized name of the sender \cr
     3) A column to indicate the name of the sender \cr
     4) A column containing the raw message \cr
     5) A column containgin a "flat" message, stripped of emoji, numbers, special characters, file attachments, sent Locations etc. \cr
     6) A column containing a tokenized version of the flat message \cr
     7) A column containing only URLs that were contained in the messages (optional: can be shortened to domains) \cr
     8) A column containing only the names of attached media files \cr
     9) A column containing only sent locations and indicators for shared live locations \cr
     10) A column containing only emoji that were used in the message \cr
     11) A column containing only extual descriptions of emoji that were used in the message \cr
     12) A column containing only emoticons (e.g. ":-)") that were used in the message \cr
     13) A column containing the number of tokens per message, derived from the "flattened" message \cr
     14) A column containing WhatsApp System Messages in group chats (e.g."You added Frank to the group") \cr
     15) A column specifying the order of the rows according to the timestamp the messages have on the phone used for extracting the chatlog \cr
     16) A column for specifying the order of the rows as they are displayed on the phone used for extracting the chatlog \cr
}
\description{
Creates a dataframe from an exported WhatsApp textfile containing one row per message. Some columns
are saved as lists using the I() function so that multiple elements can be stored per message, while still maintaining
the general structure of one row per message. These columns should be trated as lists or unlisted first.
}
\examples{
data <- parse_chat(system.file("englishandroid24h.txt", package = "WhatsR"))
}
