% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lexical_dispersion.R
\name{plot_lexical_dispersion}
\alias{plot_lexical_dispersion}
\title{Lexical Disperson plots for Keywords in WhatsApp chatlogs}
\usage{
plot_lexical_dispersion(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  keywords = c("hello", "world"),
  return.data = FALSE,
  excludeSM = FALSE,
  ...
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with code{\link[WhatsR]{parse_chat}}}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{keywords}{A vector of keywords to be displayed, default is c("hello",world")}

\item{return.data}{Default is FALSE, returns dataframe of plot when TRUE.}

\item{excludeSM}{If TRUE, excludes the WhatsApp System Messages from the descriptive statistics. Default is FALSE.}

\item{...}{Further arguments passed down to code{\link[qdap]{dispersion_plot}}}
}
\value{
Lexical Dispersion plots for specified keywords
}
\description{
Visualizes the occurance of specific keywords within the chat
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_lexical_dispersion(data, keywords = c("auch"))
}
