% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_messages.R
\name{plot_messages}
\alias{plot_messages}
\title{Amount of messages per person}
\usage{
plot_messages(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  plot = "bar",
  return.data = TRUE
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with code{\link[WhatsR]{parse_chat}}}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{plot}{Type of plot to be returned, options include "bar" and "pie". Default is "bar".}

\item{return.data}{If TRUE, returns the subsetted dataframe. Default is FALSE}
}
\value{
Plots summarizing the amount of messages per person
}
\description{
Plots summarizing the amount of messages per person
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_messages(data)
}
