% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tokens_over_time.R
\name{plot_tokens_over_time}
\alias{plot_tokens_over_time}
\title{Distribution of Tokens over time}
\usage{
plot_tokens_over_time(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time(),
  plot = "alltime",
  return.data = FALSE,
  excludeSM = FALSE
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with code{\link[WhatsR]{parse_chat}}}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with code{\link[anytime]{anytime}}. Standard format is "yyyy-mm-dd hh:mm".}

\item{plot}{Time resolution for plots. Options include "year", "day", "hour", "heatmap" and "alltime". Default is "alltime".}

\item{return.data}{If TRUE, returns the subsetted dataframe. Default is FALSE.}

\item{excludeSM}{If TRUE, excludes the WhatsApp System Messages from the descriptive statistics. Default is FALSE.}
}
\value{
A summary of tokens over time
}
\description{
Summarizes the distribution of tokens over time
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
plot_tokens_over_time(data)
}
