% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_chat.R
\name{summarize_chat}
\alias{summarize_chat}
\title{Basic WhatsApp Chatlog Statistics}
\usage{
summarize_chat(data, excludeSM = TRUE)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with code{\link[WhatsR]{parse_chat}}}

\item{excludeSM}{If TRUE, excludes the WhatsAppSystemmessages from the descriptive statistics. Default is TRUE.}
}
\value{
A list containing:

     1) The number of participants in the chat \cr
     2) The Author of the first message \cr
     3) The body of the first message \cr
     4) The date of the first message and the last message \cr
     6) The total duration of the chat
}
\description{
Creates a list of basic information about a single WhatsApp chatlog
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
summarize_chat(data)
}
