% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_tokens_per_person.R
\name{summarize_tokens_per_person}
\alias{summarize_tokens_per_person}
\title{Token Distributions per Message}
\usage{
summarize_tokens_per_person(
  data,
  names = "all",
  starttime = anytime("1960-01-01 00:00"),
  endtime = Sys.time()
)
}
\arguments{
\item{data}{A WhatsApp chatlog that was parsed with WhatsAppParse()}

\item{names}{A vector of author names that the Plots will be restricted to}

\item{starttime}{Datetime that is used as the minimum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}

\item{endtime}{Datetime that is used as the maximum boundary for exclusion. Is parsed with anytime(). Standard format is "yyyy-mm-dd hh:mm".}
}
\value{
A summary of tokens per message disitribution per author
}
\description{
Summarizing the distribution of tokens per message
}
\examples{
data <- readRDS(system.file("ParsedWhatsAppChat.rds", package = "WhatsR"))
summarize_tokens_per_person(data)
}
